refactor(convert): move pseudo inject_inline_pseudo_images tests into home module
